Change require to placeholder
Uploaded code where require statement is specific to my game's implementation of PromiseSetCore (also where it was written). Changed require to a placeholder statement to fit other projects' needs. PromiseSetCore is a module, so it should also be represented as a ModuleScript in code bases.

Promise library needs to be branched to ReplicatedFirst if PromiseSetCore is used there, otherwise it may not have access to Promises and thus render this code useless. It will, however, run as soon as Promises are replicated if you wait for the resource to exist.