Update and rename Samples/OOP Inheritance.lua to Samples/Inheritance/Game.lua
Trying out better formatting and new standards when working with OOP in Lua. Superior to previous sample under the loss of being able to copy the whole source raw and test. More work will be required to piece the super and sub class modules together to test and understand what's happening, especially for novices.